Display enemy health and keep enemies in battle
