+++ main.go
@@ -25,10 +25,11 @@ func init() {
 func (g *Game) Draw(screen *ebiten.Image) {
 	ui.CurrentMenu.Draw(screen)
 	ui.DrawPlayerInformation(screen)
+	ui.DrawEnemyInformation(screen)
 }
 
 func (g *Game) Layout(outsideWidth, outsideHeight int) (screenWidth, screenHeight int) {
-	return 320, 240
+	return 800, 600
 }
 
 func main() {
